@@ -8,24 +8,27 @@
   "author": "Christopher Hiller <boneskull@boneskull.com> (https://boneskull.com/)",
   "license": "MIT",
   "dependencies": {
-    "@std/esm": "^0.2.1",
-    "debug": "^3.0.0",
+    "@std/esm": "^0.18.0",
+    "common-tags": "^1.5.1",
+    "debug": "^3.1.0",
     "dotenv": "^4.0.0",
     "lodash": "^4.17.4",
+    "mkdirp": "^0.5.1",
+    "pronounceable": "^1.0.1",
     "random-js": "^1.0.8",
-    "rxjs": "^5.4.2",
+    "rxjs": "^5.5.5",
     "statistics": "^3.3.0",
     "stochator": "^0.6.2",
     "twit": "^2.2.9",
-    "yargs": "^8.0.2"
+    "yargs": "^10.0.3"
   },
   "devDependencies": {
-    "eslint": "^3.19.0",
+    "eslint": "^4.13.1",
     "eslint-config-semistandard": "^11.0.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-node": "^4.2.2",
-    "eslint-plugin-promise": "^3.5.0",
+    "eslint-config-standard": "^11.0.0-beta.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0",
     "eslint-plugin-standard": "^3.0.1"
   },
   "repository": {
