@@ -1,10 +1,7 @@
 import yargs from 'yargs';
-import {main} from './main';
+import {join} from 'path';
 
-const argv = yargs.option('dry-run', {
-  default: false,
-  description: 'Do everything except post to Twitter'
-})
-  .help().argv;
-
-main(argv);
+yargs.usage('$0 [command]')
+  .commandDir(join(__dirname, 'commands'))
+  .demandCommand(1, 1)
+  .help().parse();
