@@ -1,27 +1,21 @@
 import _ from 'lodash/fp';
-import tuples from '../data/letter-freq-tuples.json';
-import triples from '../data/letter-freq-triples.json';
 import d from 'debug';
 
 const DBL_CONSONANT_BLACKLIST = 'CHJKQVWXYZ';
 
 const debug = d('bobsonbot:filters');
 
-function noDoublesAtStart (word) {
-  return !(word.slice(0, 2) !== 'AA' && /^([\w])\1/.test(word));
-}
+const doubleConsonantRegex = new RegExp(`([${DBL_CONSONANT_BLACKLIST}])\\1`);
 
-function noDoubleConsonants (word) {
-  return !new RegExp(`([${DBL_CONSONANT_BLACKLIST}])\\1`).test(word);
-}
+const noDoublesAtStart = word => !(word.slice(0, 2) !== 'AA' &&
+  /^([\w])\1/.test(word));
 
-function wordScore (word) {
+const noDoubleConsonants = word => !doubleConsonantRegex.test(word);
+
+const wordScore = ({tuples, triples}, word) => {
   let scores = [];
   word = word.toLowerCase();
   switch (word.length) {
-    case 1:
-      scores.push(1);
-      break;
     case 2:
       for (let i = 0; i < word.length - 1; i++) {
         scores.push(_.getOr(0, `['${word[i]}']['${word[i + 1]}']`, tuples));
@@ -35,14 +29,15 @@ function wordScore (word) {
       }
       break;
   }
-  let i = scores.length;
+  // let i = scores.length;
   // while (i--) {
   //   if (scores[i] < 0.002) {
   //     return 0;
   //   }
   // }
+
   return _.mean(scores);
-}
+};
 
 const isPronounceable = _.curry((statMap, word) => {
   // fix irishify and scottishify
